QuantumCircuitOpt.jl Function References

QuantumCircuitOpt.GateData — Type

GateData

The composite mutable struct, GateData, type of the gate, the complex matrix form of the gate, full sized real form of the gate, inverse of the gate and a boolean which states if the gate has all real entries.

QuantumCircuitOpt.QuantumCircuitModel — Type

QuantumCircuitModel

The composite mutable struct, QuantumCircuitModel, holds dictionaries for input data, abstract JuMP model for optimization, variable references and result from solving the JuMP model.

QuantumCircuitOpt.C2SXGate — Method

C2SXGate()

Two-qubit controlled (SXGate), which is also the square root of the Pauli-XGate.

Circuit Representation

q_0: ─────■─────
     ┌────┴────┐
q_1: ┤ sqrt(X) ├
     └─────────┘

Matrix Representation

\[C2SXGate = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0.5+0.5i & 0.5-0.5i \\ 0 & 0 & 0.5-0.5i & 0.5+0.5i \end{pmatrix}\]

QuantumCircuitOpt.CCZGate — Method

CCZGate()

Three-qubit controlled-controlled Z gate.

Circuit Representation

q_0: ─■─
      │
q_1: ─■─
      │
q_2: ─■─

Matrix Representation

\[CCZGate = \begin{pmatrix} 1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 1 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 1 & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 0 & -1 \\ \end{pmatrix}\]

QuantumCircuitOpt.CHGate — Method

CHGate()

Two-qubit, symmetric, controlled Hadamard gate (HGate).

Circuit Representation

q_0: ──■──
     ┌─┴─┐  
q_1: ┤ H ├    
     └───┘

Matrix Representation

\[CH = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \frac{1}{\sqrt{2}} & \frac{1}{\sqrt{2}} \\ 0 & 0 & \frac{1}{\sqrt{2}} & -\frac{1}{\sqrt{2}} \end{pmatrix}\]

QuantumCircuitOpt.CNotGate — Method

CNotGate()

Two-qubit controlled NOT gate with control and target on first and second qubits, respectively. This is also called the controlled X gate (CXGate).

Circuit Representation

q_0: ──■──
     ┌─┴─┐
q_1: ┤ X ├
     └───┘

Matrix Representation

\[CNot = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 0 & 1 & 0 \end{pmatrix}\]

QuantumCircuitOpt.CNotRevGate — Method

CNotRevGate()

Two-qubit reverse controlled NOT gate, with target and control on first and second qubits, respectively.

Circuit Representation

     ┌───┐
q_0: ┤ X ├
     └─┬─┘
q_1: ──■──

Matrix Representation

\[CNotRev = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 0 & 1 & 0 \\ 0 & 1 & 0 & 0 \end{pmatrix}\]

QuantumCircuitOpt.CRXGate — Method

CRXGate(θ::Number)

Two-qubit controlled RXGate.

Circuit Representation

q_0: ────■────
     ┌───┴───┐
q_1: ┤ RX(ϴ) ├
     └───────┘

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} CRX(\theta)\ q_1, q_0 = |0\rangle\langle0| \otimes I + |1\rangle\langle1| \otimes RX(\theta) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \cos{\th} & -i\sin{\th} \\ 0 & 0 & -i\sin{\th} & \cos{\th} \end{pmatrix}\]

QuantumCircuitOpt.CRYGate — Method

CRYGate(θ::Number)

Two-qubit controlled RYGate.

Circuit Representation

q_0: ────■────
     ┌───┴───┐
q_1: ┤ RY(ϴ) ├
     └───────┘

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} CRY(\theta)\ q_1, q_0 = |0\rangle\langle0| \otimes I + |1\rangle\langle1| \otimes RY(\theta) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \cos{\th} & -\sin{\th} \\ 0 & 0 & \sin{\th} & \cos{\th} \end{pmatrix}\]

QuantumCircuitOpt.CRZGate — Method

CRZGate(θ::Number)

Two-qubit controlled RZGate.

Circuit Representation

q_0: ────■────
     ┌───┴───┐
q_1: ┤ RZ(ϴ) ├
     └───────┘

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} CRZ(\theta)\ q_1, q_0 = |0\rangle\langle0| \otimes I + |1\rangle\langle1| \otimes RZ(\theta) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & e^{-i\th} & 0 \\ 0 & 0 & 0 & e^{i\th} \end{pmatrix}\]

QuantumCircuitOpt.CSwapGate — Method

CSwapGate()

Three-qubit, controlled SwapGate, also known as the Fredkin gate.

Circuit Representation

q_0: ─■─
      │
q_1: ─X─
      │
q_2: ─X─

Matrix Representation

\[CSwapGate = \begin{pmatrix} 1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 1 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 1 & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & 0 & 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1 \\ \end{pmatrix}\]

QuantumCircuitOpt.CU3Gate — Method

CU3Gate(θ::Number, ϕ::Number, λ::Number)

Two-qubit, controlled version of the universal rotation gate with three Euler angles (U3Gate).

Circuit Representation

q_0: ──────■──────
     ┌─────┴─────┐
q_1: ┤ U3(ϴ,φ,λ) ├
     └───────────┘

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} CU3(\theta, \phi, \lambda)\ q_1, q_0 = |0\rangle\langle 0| \otimes I + |1\rangle\langle 1| \otimes U3(\theta,\phi,\lambda) = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & \cos(\th) & -e^{i\lambda}\sin(\th) \\ 0 & 0 & e^{i\phi}\sin(\th) & e^{i(\phi+\lambda)}\cos(\th) \end{pmatrix}\]

QuantumCircuitOpt.CVGate — Method

CVGate()

Two-qubit, controlled-V gate, which is also the same as Controlled square-root of X gate (C2SXGate).

Circuit Representation

q_0: ──■──     
     ┌─┴─┐    
q_1: ┤ V ├     
     └───┘

Matrix Representation

\[CV = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0.5+0.5i & 0.5-0.5i \\ 0 & 0 & 0.5-0.5i & 0.5+0.5i \end{pmatrix}\]

QuantumCircuitOpt.CXGate — Method

CXGate()

Two-qubit controlled XGate, which is also the same as CNotGate.

Circuit Representation

q_0: ──■──
     ┌─┴─┐
q_1: ┤ X ├
     └───┘

Matrix Representation

\[CX = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 0 & 1 & 0 \end{pmatrix}\]

QuantumCircuitOpt.CYGate — Method

CYGate()

Two-qubit controlled YGate.

Circuit Representation

q_0: ──■──
     ┌─┴─┐
q_1: ┤ Y ├
     └───┘

Matrix Representation

\[CY = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & -i \\ 0 & 0 & i & 0 \end{pmatrix}\]

QuantumCircuitOpt.CZGate — Method

CZGate()

Two-qubit, symmetric, controlled ZGate.

Circuit Representation

q_0: ──■──     ─■─
     ┌─┴─┐  ≡   │
q_1: ┤ Z ├     ─■─
     └───┘

Matrix Representation

\[CZ = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & -1 \end{pmatrix}\]

QuantumCircuitOpt.DCXGate — Method

DCXGate()

Two-qubit double controlled NOT gate consisting of two back-to-back CNotGates with alternate controls.

Circuit Representation

          ┌───┐
q_0: ──■──┤ X ├
     ┌─┴─┐└─┬─┘
q_1: ┤ X ├──■──
     └───┘

Matrix Representation

\[DCNot = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \end{pmatrix}\]

QuantumCircuitOpt.HCoinGate — Method

HCoinGate()

Two-qubit, Hadamard Coin gate when implemented in tune with the quantum cellular automata. Reference: https://doi.org/10.1007/s11128-018-1983-x, https://arxiv.org/pdf/2106.03115.pdf

Matrix Representation

\[HCoinGate = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & \frac{1}{\sqrt{2}} & \frac{1}{\sqrt{2}} & 0 \\ 0 & \frac{1}{\sqrt{2}} & -\frac{1}{\sqrt{2}} & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

QuantumCircuitOpt.HGate — Method

HGate(num_qubits::Int64)

Single-qubit Hadamard gate, which is a $\pi$ rotation about the X+Z axis, thus equivalent to U3Gate($\frac{\pi}{2},0,\pi$)

Matrix Representation

\[H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & -1 \end{pmatrix}\]

QuantumCircuitOpt.IGate — Method

IGate(num_qubits::Int64)

Identity gate corresponds to the regular identity matrix for a given number of qubits.

Matrix Representation (num_qubits = 1)

\[I = \begin{pmatrix} 1 & 0 \\ 0 & 1 \end{pmatrix}\]

QuantumCircuitOpt.MGate — Method

MGate()

Two-qubit Magic gate, also known as the Ising coupling or the XX gate.

Reference: https://doi.org/10.1103/PhysRevA.69.032315

Circuit Representation

      ┌───┐        ┌───┐
q_0: ─┤ X ├────────┤ S ├
      └─┬─┘        └─┬─┘        
        │   ┌───┐  ┌─┴─┐
q_1: ───■───┤ H ├──┤ S ├
            └───┘  └───┘

Matrix Representation

\[M = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & i & 0 & 0 \\ 0 & 0 & i & 1 \\ 0 & 0 & i & -1 \\ 1 & -i & 0 & 0 \end{pmatrix}\]

QuantumCircuitOpt.NegIGate — Method

NegIGate()

Two-qubit negative Identity gate.

Circuit Representation

q_1: ┤ RZ(ϴ) ├
     └───────┘
q_0: ─────────

Matrix Representation

\[NegI = \begin{pmatrix} -1 & 0 & 0 & 0 \\ 0 & -1 & 0 & 0 \\ 0 & 0 & -1 & 0 \\ 0 & 0 & 0 & -1 \end{pmatrix}\]

QuantumCircuitOpt.PeresGate — Method

PeresGate()

Three-qubit Peres gate. This gate is equivalent to ToffoliGate followed by the CNotGate in 3 qubits. Reference: https://doi.org/10.1103/PhysRevA.32.3266

Circuit Representation

q_0: ──■─────■──          
       │   ┌─┴─┐
q_1: ──■───┤ X ├
     ┌─┴─┐ └───┘
q_2: ┤ X ├──────
     └───┘

Matrix Representation

\[PeresGate = \begin{pmatrix} 1 & 0 & 0 & 0 & 0 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 1 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1 \\ 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & 0 & 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 0 & 1 & 0 & 0 & 0 \\ \end{pmatrix}\]

QuantumCircuitOpt.PhaseGate — Method

PhaseGate()

Single-qubit rotation gate about the Z axis. This is also equivalent to U3Gate($0,0,\lambda$). This gate is also referred to as the U1Gate.

Matrix Representation

\[P(\lambda) = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\lambda} \end{pmatrix}\]

QuantumCircuitOpt.QFT2Gate — Method

QFT2Gate()

Two-qubit Quantum Fourier Transform (QFT) gate, where the QFT operation on n-qubits is given by:

\[|j\rangle \mapsto \frac{1}{2^{n/2}} \sum_{k=0}^{2^n - 1} e^{2\pi ijk / 2^n} |k\rangle\]

Circuit Representation

     ┌──────┐
q_0: ┤      ├
     │ QFT2 │   
q_1: ┤      ├ 
     └──────┘ 

Matrix Representation

\[M = \frac{1}{2} \begin{pmatrix} 1 & 1 & 1 & 1 \\ 1 & i & -1 & -i \\ 1 & -1 & 1 & -1 \\ 1 & -i & -1 & i \end{pmatrix}\]

QuantumCircuitOpt.RXGate — Method

RXGate(θ::Number)

A single-qubit Pauli gate which represents rotation about the X axis.

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} RX(\theta) = exp(-i \th X) = \begin{pmatrix} \cos{\th} & -i\sin{\th} \\ -i\sin{\th} & \cos{\th} \end{pmatrix}\]

QuantumCircuitOpt.RYGate — Method

RYGate(θ::Number)

A single-qubit Pauli gate which represents rotation about the Y axis.

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} RY(\theta) = exp(-i \th Y) = \begin{pmatrix} \cos{\th} & -\sin{\th} \\ \sin{\th} & \cos{\th} \end{pmatrix}\]

QuantumCircuitOpt.RZGate — Method

RZGate(θ::Number)

A single-qubit Pauli gate which represents rotation about the Z axis. This gate is also equivalent to U1Gate up to a phase factor, that is, $RZ(\theta) = e^{-i{\theta}/2}U1(\theta)$.

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} RZ(\theta) = exp(-i\th Z) = \begin{pmatrix} e^{-i\th} & 0 \\ 0 & e^{i\th} \end{pmatrix}\]

QuantumCircuitOpt.SGate — Method

SGate()

Single-qubit S gate, equivalent to U3Gate($0,0,\frac{\pi}{2}$). This gate is also referred to as a Clifford gate, P gate or a square-root of Pauli-ZGate. Historically, this is also called as the phase gate (denoted by P), since it shifts the phase of the one state relative to the zero state.

Matrix Representation

\[S = \begin{pmatrix} 1 & 0 \\ 0 & i \end{pmatrix}\]

QuantumCircuitOpt.SXGate — Method

SXGate()

Single-qubit square root of pauli-XGate.

Matrix Representation

\[\sqrt{X} = \frac{1}{2} \begin{pmatrix} 1 + i & 1 - i \\ 1 - i & 1 + i \end{pmatrix}\]

QuantumCircuitOpt.SXdaggerGate — Method

SXdaggerGate()

Single-qubit hermitian conjugate of the square root of pauli-XGate, or the SXGate.

Matrix Representation

\[\sqrt{X}^{\dagger} = \frac{1}{2} \begin{pmatrix} 1 - i & 1 + i \\ 1 + i & 1 - i \end{pmatrix}\]

QuantumCircuitOpt.SdaggerGate — Method

SdaggerGate()

Single-qubit, hermitian conjugate of the SGate. This is also an alternative square root of the ZGate.

Matrix Representation

\[S = \begin{pmatrix} 1 & 0 \\ 0 & -i \end{pmatrix}\]

QuantumCircuitOpt.SwapGate — Method

SwapGate()

Two-qubit, symmetric, SWAP gate.

Circuit Representation

q_0: ─X─
      │
q_1: ─X─

Matrix Representation

\[SWAP = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

QuantumCircuitOpt.TGate — Method

TGate()

Single-qubit T gate, equivalent to U3Gate($0,0,\frac{\pi}{4}$). This gate is also referred to as a $\frac{\pi}{8}$ gate or as a fourth-root of Pauli-ZGate.

Matrix Representation

\[T = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\pi/4} \end{pmatrix}\]

QuantumCircuitOpt.TdaggerGate — Method

TdaggerGate()

Single-qubit, hermitian conjugate of the TGate. This gate is equivalent to U3Gate($0,0,-\frac{\pi}{4}$). This gate is also referred to as the fourth-root of Pauli-ZGate.

Matrix Representation

\[T^{\dagger} = \begin{pmatrix} 1 & 0 \\ 0 & e^{-i\pi/4} \end{pmatrix}\]

QuantumCircuitOpt.ToffoliGate — Method

ToffoliGate()

Three-qubit Toffoli gate, also known as the CCX (controlled-controlled-NOT) gate.

Circuit Representation

q_0: ──■──
       │
q_1: ──■──
     ┌─┴─┐
q_2: ┤ X ├
     └───┘

Matrix Representation

\[Toffoli = |0 \rangle \langle 0| \otimes I \otimes I + |1 \rangle \langle 1| \otimes CXGate = \begin{pmatrix} 1 & 0 & 0 & 0 & 0 & 0 & 0 & 0\\ 0 & 1 & 0 & 0 & 0 & 0 & 0 & 0\\ 0 & 0 & 1 & 0 & 0 & 0 & 0 & 0\\ 0 & 0 & 0 & 1 & 0 & 0 & 0 & 0\\ 0 & 0 & 0 & 0 & 1 & 0 & 0 & 0\\ 0 & 0 & 0 & 0 & 0 & 1 & 0 & 0\\ 0 & 0 & 0 & 0 & 0 & 0 & 0 & 1\\ 0 & 0 & 0 & 0 & 0 & 0 & 1 & 0 \end{pmatrix}\]

QuantumCircuitOpt.U1Gate — Method

U1Gate(λ::Number)

Universal single-qubit rotation gate with one Euler angle, $\lambda$. U1Gate represents rotation about the Z axis and is the special case of U3Gate. Also note that $U1(\pi) =$ZGate, $U1(\pi/2) =$SGate and $U1(\pi/4) =$TGate.

Matrix Representation

\[U1(\lambda) = \begin{pmatrix} 1 & 0 \\ 0 & e^{i\lambda} \end{pmatrix}\]

QuantumCircuitOpt.U2Gate — Method

U2Gate(ϕ::Number, λ::Number)

Universal single-qubit rotation gate with two Euler angles, $\phi$ and $\lambda$. U2Gate is the special case of U3Gate.

Matrix Representation

\[U2(\phi, \lambda) = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & -e^{i\lambda} \\ e^{i\phi} & e^{i(\phi+\lambda)} \end{pmatrix}\]

QuantumCircuitOpt.U3Gate — Method

U3Gate(θ::Number, ϕ::Number, λ::Number)

Universal single-qubit rotation gate with three Euler angles, $\theta$, $\phi$ and $\lambda$.

Matrix Representation

\[\newcommand{\th}{\frac{\theta}{2}} U3(\theta, \phi, \lambda) = \begin{pmatrix} \cos(\th) & -e^{i\lambda}\sin(\th) \\ e^{i\phi}\sin(\th) & e^{i(\phi+\lambda)}\cos(\th) \end{pmatrix}\]

QuantumCircuitOpt.WGate — Method

WGate()

Two-qubit, W hermitian gate, typically useful to diagonlize the (SwapGate).

Matrix Representation

\[CV = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & \frac{1}{\sqrt{2}} & \frac{1}{\sqrt{2}} & 0 \\ 0 & \frac{1}{\sqrt{2}} & -\frac{1}{\sqrt{2}} & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

QuantumCircuitOpt.XGate — Method

XGate()

Single-qubit Pauli-X gate ($\sigma_x$), equivalent to U3Gate($\pi,0,\pi$)

Matrix Representation

\[X = \begin{pmatrix} 0 & 1 \\ 1 & 0 \end{pmatrix}\]

QuantumCircuitOpt.YGate — Method

YGate()

Single-qubit Pauli-Y gate ($\sigma_y$), equivalent to U3Gate($\pi,\frac{\pi}{2},\frac{\pi}{2}$)

Matrix Representation

\[Y = \begin{pmatrix} 0 & -i \\ i & 0 \end{pmatrix}\]

QuantumCircuitOpt.ZGate — Method

ZGate()

Single-qubit Pauli-Z gate ($\sigma_z$), equivalent to U3Gate($0,0,\pi$)

Matrix Representation

\[Z = \begin{pmatrix} 1 & 0 \\ 0 & -1 \end{pmatrix}\]

QuantumCircuitOpt.build_QCModel_result — Method

QuantumCircuitOpt.complex_to_real_matrix — Method

complex_to_real_matrix(M::Array{Complex{Float64},2})

Given a complex-valued 2D matrix of size NxN, complextoreal_matrix function returns a real-valued matrix of size 2Nx2N.

QuantumCircuitOpt.eliminate_nonunique_gates — Method

eliminate_nonunique_gates(gates_dict::Dict{String, Any})

QuantumCircuitOpt.gate_element_bounds — Method

gate_element_bounds(M::Array{Float64,3})

Given a set of elementary gates, {G1, G2, ... Gn}, `gateelement_bounds` function evaluates the range of every co-ordinate of the superimposed gates, over all possible gates.

QuantumCircuitOpt.get_commutative_gate_pairs — Method

get_commutative_gate_pairs(M::Dict{String,Any}; identity_pairs = true)

Given a dictionary of gates, as processed in src/data.jl, get_commutative_gate_pairs outputs all pairs of commuting matrices. Optional argument, identity_pairs can be set to false if identity matrix need not be part of the commuting pairs.

QuantumCircuitOpt.get_compressed_solutions — Method

getcompressedsolutions returns sequence of gates after compressing adjacent pair of gates represented on two separate qubits. For example, gates H1 and H2 appearing in a sequence will be compressed to H1⊗H2.

QuantumCircuitOpt.get_full_sized_gate — Method

get_full_sized_gate(input::String, num_qubits::Int64; matrix = nothing, qubit_location = nothing)

For a given string and number of qubits in the input specified input, this function returns a full sized gate with respect to the input number of qubits.

QuantumCircuitOpt.get_objective_bound — Method

QuantumCircuitOpt.get_objective_value — Method

QuantumCircuitOpt.get_quantum_gates — Method

get_quantum_gates(params::Dict{String, Any}, elementary_gates::Array{String,1})

Given a vector of input with the names of gates (see examples folder), get_quantum_gates function returns the corresponding elementary gates in the three-dimensional complex matrix form.

QuantumCircuitOpt.iSwapGate — Method

iSwapGate()

Two-qubit, symmetric and clifford, iSWAP gate.

Circuit Representation

q_0: ─⨂─
      │     
q_1: ─⨂─    

Minimum depth representation

      ┌───┐     ┌───┐ ┌───┐
q_0: ─┤ X ├──■──┤ S ├─┤ X ├─
      └─┬─┘┌─┴─┐└───┘ └─┬─┘
q_1: ───■──┤ X ├────────■──
           └───┘

Matrix Representation

\[iSWAP = \begin{pmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & i & 0 \\ 0 & i & 0 & 0 \\ 0 & 0 & 0 & 1 \end{pmatrix}\]

QuantumCircuitOpt.kron_single_gate — Method

kron_single_gate(num_qubits::Int64, M::Array{Complex{Float64},2}, qubit_loc::String)

Given number of qubits of the circuit, the complex-valued gate and the qubit location ("q1","q2',"q3",...), kronsinglegate function returns a full-sized gate after applying appropriate kronecker products.

QuantumCircuitOpt.logger_config! — Method

allows the user to set the logging level without the need to add Memento

QuantumCircuitOpt.optimize_QCModel! — Method

QuantumCircuitOpt.real_to_complex_matrix — Method

real_to_complex_matrix(M::Array{Complex{Float64},2})

Given a real-valued 2D matrix of size 2Nx2N, realtocomplex_matrix function returns a complex-valued matrix of size NxN, if the input matrix is in a valid complex matrix form.

QuantumCircuitOpt.relaxation_bilinear — Method

relaxation_bilinear(m::JuMP.Model, xy::JuMP.VariableRef, x::JuMP.VariableRef, y::JuMP.VariableRef)

general relaxation of binlinear term (McCormick), which can be used to obtain specific variants in partiuclar cases of variables (like binary)

z >= JuMP.lower_bound(x)*y + JuMP.lower_bound(y)*x - JuMP.lower_bound(x)*JuMP.lower_bound(y)
z >= JuMP.upper_bound(x)*y + JuMP.upper_bound(y)*x - JuMP.upper_bound(x)*JuMP.upper_bound(y)
z <= JuMP.lower_bound(x)*y + JuMP.upper_bound(y)*x - JuMP.lower_bound(x)*JuMP.upper_bound(y)
z <= JuMP.upper_bound(x)*y + JuMP.lower_bound(y)*x - JuMP.upper_bound(x)*JuMP.lower_bound(y)

QuantumCircuitOpt.round_complex_values — Method

round_complex_values(M::Array{Complex{Float64},2})

Given a complex-valued 2D matrix, roundcomplexvalues function returns a complex-valued matrix which rounds the valuest closest to 0 and 1. This is useful to avoid numerical issues.

QuantumCircuitOpt.silence — Method

Suppresses information and warning messages output by QuantumCircuitOpt, for fine grained control use the Memento package

QuantumCircuitOpt.unique_idx — Method

unique_idx(x::AbstractArray{T})

unique_idx returns the indices of unique elements in a given array of scalar or vector inputs. Overall, this function computes faster than Julia's built-in findfirst command.

QuantumCircuitOpt.unique_matrices — Method

unique_matrices(M::Array{Float64, 3})

unique_matrices returns the unique set of matrices and the corresponding indices of unique matrices from the given set of matrices.

QuantumCircuitOpt.validate_solutions — Method

validate_solutions validates the decomposition if it is indeed exact with respect to the specified target gate.

QuantumCircuitOpt.variable_domain — Method

variable_domain(var::JuMP.VariableRef)

Computes the valid domain of a given JuMP variable taking into account bounds and the varaible's implicit bounds (e.g. binary).