Quantum Circuits Library: Multi-qubit gates

GRGate

QuantumCircuitOpt.GRGate — Function

GRGate(num_qubits::Int64, θ::Number, ϕ::Number)

A multi-qubit rotation gate with two Euler angles, $\theta$ and $\phi$, applied about the $\cos(\phi)x + \sin(\phi)y$ axis and parametrized by the number of qubits. This gate can be applied to multiple qubits simultaneously, for a given depth. The global R gate is native to atomic systems. In the one-qubit case, this gate is equivalent to the RGate.

Reference: Qiskit's circuit library

Circuit Representation (in 3 qubits)

     ┌──────────┐
q_0: ┤0         ├ 
     │          │
q_1: ┤1 GR(ϴ,φ) ├    
     │          │
q_2: ┤2         ├
     └──────────┘

Matrix Representation (in 3 qubits)

\[GR(\theta, \phi) = \exp \left(-i \sum_{i=1}^{3} (\cos(\phi)X_i + \sin(\phi)Y_i) \theta/2 \right) \\ = R(\theta, \phi) \otimes R(\theta, \phi) \otimes R(\theta, \phi) \]